+++ blocklets/ai-studio/api/src/store/migrations/20240314105359-add-datastore.ts
@@ -1,0 +1,33 @@
+import { DataTypes } from 'sequelize';
+
+import type { Migration } from '../migrate';
+
+export const up: Migration = async ({ context: queryInterface }) => {
+  await queryInterface.createTable('Datastores', {
+    id: {
+      type: DataTypes.STRING,
+      primaryKey: true,
+      allowNull: false,
+    },
+    data: {
+      type: DataTypes.JSON,
+    },
+    userId: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    sessionId: {
+      type: DataTypes.STRING,
+    },
+    createdAt: {
+      type: DataTypes.DATE,
+    },
+    updatedAt: {
+      type: DataTypes.DATE,
+    },
+  });
+};
+
+export const down: Migration = async ({ context: queryInterface }) => {
+  await queryInterface.dropTable('Datastores');
+};