+++ blocklets/ai-studio/src/components/file-editor/execute-block.tsx
@@ -204,12 +204,7 @@ export default function ExecuteBlockForm({
                   readOnly={readOnly}
                   projectId={projectId}
                   gitRef={gitRef}
-                  ContentProps={{
-                    sx: {
-                      px: 1,
-                      py: 0.5,
-                    },
-                  }}
+                  ContentProps={{ sx: { px: 1, py: 0.5 } }}
                   placeholder="Your select prompt"
                   path={path.concat('selectByPrompt')}
                   assistant={assistant}
@@ -797,7 +792,13 @@ export const ToolDialog = forwardRef<
                   name={`parameters.${parameter.name}`}
                   render={({ field }) => (
                     <PromptEditorField
-                      placeholder={`{{ ${parameter.name} }}`}
+                      placeholder={
+                        executeBlock.selectType === 'selectByPrompt'
+                          ? t('selectByPromptParameterPlaceholder')
+                          : assistantParameters.has(parameter.key)
+                            ? `{{ ${parameter.name} }}`
+                            : undefined
+                      }
                       value={field.value || ''}
                       projectId={projectId}
                       gitRef={gitRef}