+++ blocklets/ai-studio/src/libs/release.ts
@@ -12,14 +12,17 @@ export async function getReleases({
   projectId: string;
   projectRef?: string;
   assistantId?: string;
-}): Promise<{ releases: Release[] }> {
+}): Promise<{ releases: (Release & { paymentUnitAmount?: string })[] }> {
   return axios.get('/api/releases', { params: { projectId, projectRef, assistantId } }).then((res) => res.data);
 }
 
-export async function createRelease(input: CreateReleaseInput): Promise<Release> {
+export async function createRelease(input: CreateReleaseInput): Promise<Release & { paymentUnitAmount?: string }> {
   return axios.post('/api/releases', input).then((res) => res.data);
 }
 
-export async function updateRelease(releaseId: string, input: UpdateReleaseInput): Promise<Release> {
+export async function updateRelease(
+  releaseId: string,
+  input: UpdateReleaseInput
+): Promise<Release & { paymentUnitAmount?: string }> {
   return axios.patch(joinURL('/api/releases', releaseId), input).then((res) => res.data);
 }