+++ blocklets/ai-studio/src/pages/project/publish-view.tsx
@@ -1,4 +1,4 @@
-import { UpdateReleaseInput } from '@api/routes/release';
+import { CreateReleaseInput } from '@api/routes/release';
 import LoadingButton from '@app/components/loading/loading-button';
 import { useUploader } from '@app/contexts/uploader';
 import { getErrorMessage } from '@app/libs/api';
@@ -7,6 +7,7 @@ import { useLocaleContext } from '@arcblock/ux/lib/Locale/context';
 import RelativeTime from '@arcblock/ux/lib/RelativeTime';
 import Toast from '@arcblock/ux/lib/Toast';
 import { AssistantYjs } from '@blocklet/ai-runtime/types';
+import ComponentInstaller from '@blocklet/ui-react/lib/ComponentInstaller';
 import styled from '@emotion/styled';
 import { LaunchRounded } from '@mui/icons-material';
 import UploadIcon from '@mui/icons-material/Upload';
@@ -14,20 +15,25 @@ import {
   Box,
   FormControl,
   FormControlLabel,
+  FormLabel,
   IconButton,
+  InputAdornment,
   InputBase,
   Link,
   Radio,
   RadioGroup,
   Stack,
+  Switch,
+  TextField,
   Typography,
 } from '@mui/material';
 import { alpha, styled as muiStyled } from '@mui/material/styles';
-import { useEffect, useMemo, useState } from 'react';
+import { useEffect, useMemo } from 'react';
+import { Controller, useForm } from 'react-hook-form';
 import QRCode from 'react-qr-code';
 import { joinURL, withQuery } from 'ufo';
 
-import { useProjectState } from './state';
+import { saveButtonState, useAssistantChangesState, useProjectState } from './state';
 
 const TemplateImage = styled('img')({
   width: '100%',
@@ -36,6 +42,7 @@ const TemplateImage = styled('img')({
 
 const StyledFormControlLabel = muiStyled(FormControlLabel)({
   width: '50%',
+  maxWidth: 200,
   margin: 0,
   '& .MuiTypography-root': {
     width: '95%',
@@ -82,6 +89,27 @@ export default function PublishView({
   projectRef: string;
   assistant: AssistantYjs;
 }) {
+  return (
+    <ComponentInstaller
+      did={[
+        'z2qa6fvjmjew4pWJyTsKaWFuNoMUMyXDh5A1D',
+        'z2qaCNvKMv5GjouKdcDWexv6WqtHbpNPQDnAk',
+        'z8iZiDFg3vkkrPwsiba1TLXy3H9XHzFERsP8o',
+      ]}>
+      <PublishViewContent projectId={projectId} projectRef={projectRef} assistant={assistant} />
+    </ComponentInstaller>
+  );
+}
+
+function PublishViewContent({
+  projectId,
+  projectRef,
+  assistant,
+}: {
+  projectId: string;
+  projectRef: string;
+  assistant: AssistantYjs;
+}) {
   const { t, locale } = useLocaleContext();
   const uploaderRef = useUploader();
 
@@ -92,22 +120,18 @@ export default function PublishView({
 
   const release = releases?.find((i) => i.projectRef === projectRef && i.assistantId === assistant.id);
 
-  const [settings, setSettings] = useState<UpdateReleaseInput>({ template: 'default' });
+  const form = useForm<CreateReleaseInput>({});
 
   useEffect(() => {
-    setSettings({
+    form.reset({
       template: release?.template || 'default',
       icon: release?.icon || '',
       title: release?.title || '',
       description: release?.description || '',
+      paymentEnabled: release?.paymentEnabled,
+      paymentUnitAmount: release?.paymentUnitAmount,
     });
-  }, [release]);
-
-  const [loading, setLoading] = useState(false);
-
-  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setSettings({ ...settings, template: event.target.value as 'default' | 'blue' | 'green' | 'red' });
-  };
+  }, [form, release]);
 
   const releaseUrl = useMemo(() => {
     if (!release) return undefined;
@@ -118,44 +142,86 @@ export default function PublishView({
     );
   }, [release]);
 
+  const { disabled: hasUnsavedChanges } = useAssistantChangesState(projectId, projectRef);
+
+  const onSubmit = async (input: CreateReleaseInput) => {
+    try {
+      if (!hasUnsavedChanges) {
+        saveButtonState.getState().save?.();
+        Toast.info('Please save your prompt first');
+        return;
+      }
+
+      if (!release) {
+        await createRelease({
+          ...input,
+          projectRef,
+          projectId,
+          assistantId: assistant.id,
+        });
+        refetch();
+        Toast.success(t('publish.publishSuccess'));
+      } else {
+        await updateRelease(release.id, input);
+        refetch();
+        Toast.success(t('publish.updateSuccess'));
+      }
+    } catch (error) {
+      Toast.error(getErrorMessage(error));
+      throw error;
+    }
+  };
+
   return (
-    <Stack px={2} mt={1} py={1} gap={2} ml={1}>
+    <Stack px={2} mt={1} py={1} gap={2} ml={1} overflow="auto" component="form" onSubmit={form.handleSubmit(onSubmit)}>
       <FormControl>
         <Typography variant="subtitle2" mb={1}>
           {t('templates')}
         </Typography>
-        <RadioGroup
-          value={settings.template}
-          onChange={handleChange}
-          row
-          sx={{
-            rowGap: 1,
-          }}>
-          <StyledFormControlLabel
-            labelPlacement="top"
-            control={<Radio />}
-            value="default"
-            label={<TemplateImage src={joinURL(window?.blocklet?.prefix ?? '/', '/images/template-1.png')} alt="" />}
-          />
-          <StyledFormControlLabel
-            labelPlacement="top"
-            control={<Radio />}
-            value="blue"
-            label={<TemplateImage src={joinURL(window?.blocklet?.prefix ?? '/', '/images/template-2.png')} alt="" />}
-          />
-          <StyledFormControlLabel
-            labelPlacement="top"
-            control={<Radio />}
-            value="red"
-            label={<TemplateImage src={joinURL(window?.blocklet?.prefix ?? '/', '/images/template-3.png')} alt="" />}
-          />
-          <StyledFormControlLabel
-            labelPlacement="top"
-            control={<Radio />}
-            value="green"
-            label={<TemplateImage src={joinURL(window?.blocklet?.prefix ?? '/', '/images/template-4.png')} alt="" />}
-          />
-        </RadioGroup>
+        <Controller
+          name="template"
+          control={form.control}
+          render={({ field }) => (
+            <RadioGroup
+              row
+              sx={{ rowGap: 1 }}
+              value={field.value || ''}
+              onChange={(_, value) => field.onChange({ target: { value } })}>
+              <StyledFormControlLabel
+                labelPlacement="top"
+                control={<Radio />}
+                value="default"
+                label={
+                  <TemplateImage src={joinURL(window?.blocklet?.prefix ?? '/', '/images/template-1.png')} alt="" />
+                }
+              />
+              <StyledFormControlLabel
+                labelPlacement="top"
+                control={<Radio />}
+                value="blue"
+                label={
+                  <TemplateImage src={joinURL(window?.blocklet?.prefix ?? '/', '/images/template-2.png')} alt="" />
+                }
+              />
+              <StyledFormControlLabel
+                labelPlacement="top"
+                control={<Radio />}
+                value="red"
+                label={
+                  <TemplateImage src={joinURL(window?.blocklet?.prefix ?? '/', '/images/template-3.png')} alt="" />
+                }
+              />
+              <StyledFormControlLabel
+                labelPlacement="top"
+                control={<Radio />}
+                value="green"
+                label={
+                  <TemplateImage src={joinURL(window?.blocklet?.prefix ?? '/', '/images/template-4.png')} alt="" />
+                }
+              />
+            </RadioGroup>
+          )}
+        />
       </FormControl>
 
       <FormControl>
@@ -165,10 +231,10 @@ export default function PublishView({
         <BaseInput
           id="project-name"
           placeholder={t('publish.titlePlaceholder')}
-          value={settings.title}
-          onChange={(e) => setSettings({ ...settings, title: e.target.value })}
+          {...form.register('title', { setValueAs: (v) => v.trim() })}
         />
       </FormControl>
+
       <FormControl>
         <Typography mb={1} variant="subtitle2">
           {t('publish.description')}
@@ -181,47 +247,81 @@ export default function PublishView({
           placeholder={t('publish.descriptionPlaceholder')}
           minRows={5}
           id="description"
-          value={settings.description}
-          onChange={(e) => setSettings({ ...settings, description: e.target.value })}
+          {...form.register('description', { setValueAs: (v) => v.trim() })}
         />
       </FormControl>
 
       <Box>
         <Typography mb={1} variant="subtitle2">
-          {t('publish.favicon')}
+          {t('publish.logo')}
         </Typography>
 
-        <Box mb={0.5} width="40%">
-          <ImageContainer
-            onClick={() => {
-              // @ts-ignore
-              const uploader = uploaderRef?.current?.getUploader();
+        <Box mb={0.5} width="40%" maxWidth={120}>
+          <Controller
+            name="icon"
+            control={form.control}
+            render={({ field }) => (
+              <ImageContainer
+                onClick={() => {
+                  // @ts-ignore
+                  const uploader = uploaderRef?.current?.getUploader();
 
-              uploader?.open();
+                  uploader?.open();
 
-              uploader.onceUploadSuccess((data: any) => {
-                const { response } = data;
-                const url = response?.data?.url || response?.data?.fileUrl;
-                setSettings({ ...settings, icon: url });
-              });
-            }}>
-            {settings.icon ? (
-              <img className="upload-button" src={settings.icon} alt="" />
-            ) : (
-              <IconButton
-                className="upload-button"
-                key="uploader-trigger"
-                size="small"
-                sx={{ borderRadius: 0.5, bgcolor: 'rgba(0, 0, 0, 0.06)' }}>
-                <UploadIcon />
-              </IconButton>
+                  uploader.onceUploadSuccess((data: any) => {
+                    const { response } = data;
+                    const url = response?.data?.url || response?.data?.fileUrl;
+                    field.onChange({ target: { value: url } });
+                  });
+                }}>
+                {field.value ? (
+                  <img className="upload-button" src={field.value} alt="" />
+                ) : (
+                  <IconButton
+                    className="upload-button"
+                    key="uploader-trigger"
+                    size="small"
+                    sx={{ borderRadius: 0.5, bgcolor: 'rgba(0, 0, 0, 0.06)' }}>
+                    <UploadIcon />
+                  </IconButton>
+                )}
+              </ImageContainer>
             )}
-          </ImageContainer>
+          />
         </Box>
+      </Box>
 
-        <Typography variant="caption" color="text.secondary">
-          {t('publish.faviconDescription')}
+      <Box>
+        <Typography mb={1} variant="subtitle2">
+          {t('publish.payment')}
         </Typography>
+
+        <Stack gap={1}>
+          <Stack direction="row" gap={1} alignItems="center">
+            <FormLabel sx={{ width: 60 }}>{t('publish.enabled')}</FormLabel>
+            <Controller
+              name="paymentEnabled"
+              control={form.control}
+              rules={{ required: form.watch('paymentEnabled'), min: 0 }}
+              render={({ field }) => (
+                <FormControl>
+                  <Switch checked={field.value || false} onChange={field.onChange} />
+                </FormControl>
+              )}
+            />
+          </Stack>
+
+          {form.watch('paymentEnabled') && (
+            <Stack direction="row" gap={1} alignItems="center">
+              <FormLabel sx={{ width: 60 }}>{t('publish.price')}</FormLabel>
+              <TextField
+                {...form.register('paymentUnitAmount', { required: form.watch('paymentEnabled') })}
+                hiddenLabel
+                InputProps={{ endAdornment: <InputAdornment position="end">ABT / {t('publish.time')}</InputAdornment> }}
+              />
+            </Stack>
+          )}
+        </Stack>
       </Box>
 
       {release && releaseUrl && (
@@ -249,33 +349,7 @@ export default function PublishView({
       )}
 
       <Stack direction="row" gap={2} alignItems="center" sx={{ mt: 3 }}>
-        <LoadingButton
-          loading={loading}
-          variant="contained"
-          onClick={async () => {
-            try {
-              setLoading(true);
-              if (!release) {
-                await createRelease({
-                  ...settings,
-                  projectRef,
-                  projectId,
-                  assistantId: assistant.id,
-                });
-                refetch();
-                Toast.success(t('publish.publishSuccess'));
-              } else {
-                await updateRelease(release.id, { ...settings });
-                refetch();
-                Toast.success(t('publish.updateSuccess'));
-              }
-            } catch (error) {
-              Toast.error(getErrorMessage(error));
-              throw error;
-            } finally {
-              setLoading(false);
-            }
-          }}>
+        <LoadingButton type="submit" loading={form.formState.isSubmitting} variant="contained">
           {release ? t('publish.update') : t('publish.publishProject')}
         </LoadingButton>
       </Stack>