+++ packages/ai-runtime/src/core/assistant/index.ts
@@ -4,7 +4,7 @@ import { ReadableStream } from 'stream/web';
 import { ChatCompletionChunk, ChatCompletionInput, ImageGenerationInput } from '@blocklet/ai-kit/api/types';
 import { getBuildInDatasets } from '@blocklet/dataset-sdk';
 import { getRequest } from '@blocklet/dataset-sdk/request';
-import { getAllParameters } from '@blocklet/dataset-sdk/request/util';
+import { getAllParameters, getRequiredFields } from '@blocklet/dataset-sdk/request/util';
 import type { DatasetObject } from '@blocklet/dataset-sdk/types';
 import { call as callFunc } from '@blocklet/sdk/lib/component';
 import { env } from '@blocklet/sdk/lib/config';
@@ -780,6 +780,7 @@ async function runExecuteBlocks({
   const tasks: [ExecuteBlock, () => () => Promise<any>][] = [];
   const cache: { [key: string]: Promise<any> } = {};
   const datasets = await getBuildInDatasets();
+  const results = [];
 
   for (const executeBlock of executeBlocks) {
     const task = () => async () => {
@@ -808,7 +809,14 @@ async function runExecuteBlocks({
     tasks.push([executeBlock, task]);
   }
 
-  return Promise.all(tasks.map((i) => i[1]()().then((result) => [i[0], result] as const)));
+  // 确保日志输出顺序准确
+  for (const task of tasks) {
+    const result = await task[1]()();
+    results.push([task[0], result] as const);
+  }
+
+  return results;
+  // return Promise.all(tasks.map((i) => i[1]()().then((result) => [i[0], result] as const)));
 }
 
 async function runExecuteBlock({
@@ -858,6 +866,8 @@ async function runExecuteBlock({
     const result = (
       await Promise.all(
         tools.map(async (tool) => {
+          const currentTaskId = taskIdGenerator.nextId().toString();
+
           if (tool?.from === 'dataset') {
             const dataset = (datasets || []).find((x) => x.id === tool.id);
 
@@ -875,18 +885,42 @@ async function runExecuteBlock({
               );
 
               const params: { [key: string]: string } = {
-                sessionId: sessionId || '',
                 userId: user?.did || '',
+                sessionId: sessionId || '',
                 assistantId: assistant.id || '',
               };
+
+              const callbackParams = {
+                taskId: currentTaskId,
+                parentTaskId,
+                assistantId: assistant.id,
+                assistantName: `${executeBlock.variable ?? assistant.name}`,
+              };
+
+              callback?.({
+                type: AssistantResponseType.EXECUTE,
+                ...callbackParams,
+                execution: { currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_START },
+              });
+
+              callback?.({
+                type: AssistantResponseType.INPUT,
+                ...callbackParams,
+                inputParameters: requestData,
+              });
 
               const response = await getRequest(dataset, requestData, { user, params });
 
               callback?.({
                 type: AssistantResponseType.CHUNK,
-                taskId: taskIdGenerator.nextId().toString(),
-                assistantId: tool.id,
-                delta: { content: typeof response.data === 'string' ? response.data : JSON.stringify(response.data) },
+                ...callbackParams,
+                delta: { content: JSON.stringify(response.data) },
+              });
+
+              callback?.({
+                type: AssistantResponseType.EXECUTE,
+                ...callbackParams,
+                execution: { currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_END },
               });
 
               return response.data;
@@ -941,7 +975,7 @@ async function runExecuteBlock({
           );
 
           return runAssistant({
-            taskId: taskIdGenerator.nextId().toString(),
+            taskId: currentTaskId,
             callAI,
             callAIImage,
             getAssistant,
@@ -949,6 +983,8 @@ async function runExecuteBlock({
             parameters: args,
             parentTaskId,
             callback,
+            user,
+            sessionId,
           });
         })
       )
@@ -974,11 +1010,14 @@ async function runExecuteBlock({
             const dataset = (datasets || []).find((x) => x.id === tool.id);
             if (!dataset) return undefined;
 
+            const name = dataset.summary || dataset.description || '';
+
             const datasetParameters = getAllParameters(dataset)
               .filter((i): i is typeof i => !!i && !tool.parameters?.[i.name])
-              .map((i) => [i.name, { type: 'string', description: i.description ?? '' }]);
+              .map((i) => [i.name, { type: 'string', name, description: i.description ?? '' }]);
 
-            const name = dataset.summary || dataset.description || '';
+            const required = getRequiredFields(dataset);
+
             return {
               tool,
               toolAssistant: dataset,
@@ -988,6 +1027,7 @@ async function runExecuteBlock({
                 parameters: {
                   type: 'object',
                   properties: Object.fromEntries(datasetParameters),
+                  required: required?.length ? required : undefined,
                 },
               },
             };
@@ -1040,6 +1080,11 @@ async function runExecuteBlock({
               ];
             });
 
+          const required = (toolAssistant.parameters ?? [])
+            .filter((i): i is typeof i & { key: string } => !!i.key)
+            .filter((x) => x.required)
+            .map((x) => x.key);
+
           return {
             tool,
             toolAssistant,
@@ -1051,6 +1096,7 @@ async function runExecuteBlock({
               parameters: {
                 type: 'object',
                 properties: Object.fromEntries(toolParameters),
+                required: required?.length ? required : undefined,
               },
             },
           };
@@ -1066,6 +1112,7 @@ async function runExecuteBlock({
       modelParameters: executeBlock.executeModel,
       assistantName: `${executeBlock.variable ?? assistant.name}-select`,
     });
+
     const response = await callAI({
       assistant,
       input: {
@@ -1112,6 +1159,7 @@ async function runExecuteBlock({
     }
 
     const toolAssistantMap = Object.fromEntries(toolAssistants.map((i) => [i.function.name, i]));
+
     const result =
       calls &&
       (await Promise.all(
@@ -1121,6 +1169,7 @@ async function runExecuteBlock({
           const tool = toolAssistantMap[call.function.name];
           if (!tool) return undefined;
           const args = JSON.parse(call.function.arguments);
+          const currentTaskId = taskIdGenerator.nextId().toString();
 
           if (tool.tool.from === 'dataset') {
             const assistant = tool?.toolAssistant as DatasetObject;
@@ -1134,13 +1183,43 @@ async function runExecuteBlock({
               }) ?? []
             );
 
-            const response = await getRequest(assistant, args, { user, params: sessionId ? { sessionId } : {} });
+            const params: { [key: string]: string } = {
+              userId: user?.did || '',
+              sessionId: sessionId || '',
+              assistantId: assistant.id || '',
+            };
+
+            const callbackParams = {
+              taskId: currentTaskId,
+              parentTaskId,
+              assistantId: assistant.id,
+              assistantName: `${executeBlock.variable ?? assistant.name}`,
+            };
+
+            callback?.({
+              type: AssistantResponseType.EXECUTE,
+              ...callbackParams,
+              execution: { currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_START },
+            });
+
+            callback?.({
+              type: AssistantResponseType.INPUT,
+              ...callbackParams,
+              inputParameters: args,
+            });
+
+            const response = await getRequest(assistant, args, { user, params });
 
             callback?.({
               type: AssistantResponseType.CHUNK,
-              taskId: taskIdGenerator.nextId().toString(),
-              assistantId: assistant.id,
-              delta: { content: typeof response.data === 'string' ? response.data : JSON.stringify(response.data) },
+              ...callbackParams,
+              delta: { content: JSON.stringify(response.data) },
+            });
+
+            callback?.({
+              type: AssistantResponseType.EXECUTE,
+              ...callbackParams,
+              execution: { currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_END },
             });
 
             return response.data;
@@ -1184,7 +1263,7 @@ async function runExecuteBlock({
           );
 
           const res = await runAssistant({
-            taskId: taskIdGenerator.nextId().toString(),
+            taskId: currentTaskId,
             callAI,
             callAIImage,
             getAssistant,
@@ -1192,6 +1271,8 @@ async function runExecuteBlock({
             parameters: args,
             parentTaskId: taskId,
             callback,
+            user,
+            sessionId,
           });
 
           if (tool.tool?.onEnd === OnTaskCompletion.EXIT) {