+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,3 +1,26 @@
+## 0.1.287 (2024-3-15)
+
+- feat: add keystore db for json search
+
+## 0.1.286 (2024-3-15)
+
+- fix: require save before publish
+- fix: ignore error message when counting usage
+
+## 0.1.285 (2024-3-15)
+
+- chore: change yarn to pnpm
+
+## 0.1.284 (2024-3-13)
+
+- chore: add pages-kit component dep
+- feat: check user subscription and report usage if needed
+- feat: support config payment for release
+
+## 0.1.283 (2024-3-12)
+
+- fix: use async fs api
+
 ## 0.1.282 (2024-3-12)
 
 - feat: support history message datasets api